+++ src/server/world/paletted_container.cpp
@@ -2,17 +2,17 @@
 
 #include <utility>
 
-PalettedContainer::PalettedContainer(Palette* palette, std::vector<short> data)
+PalettedContainer::PalettedContainer(Palette* palette, std::vector<int> data)
     : palette(palette),
       data(std::move(data)) {}
 
 PalettedContainer::~PalettedContainer() = default;
 
-std::vector<short> PalettedContainer::getData() const {
+std::vector<int> PalettedContainer::getData() const {
   return data;
 }
 
-void PalettedContainer::setData(std::vector<short> v) {
+void PalettedContainer::setData(std::vector<int> v) {
   data = std::move(v);
   dirty = true;
 }
